+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-04-04 14:17:35
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-04-04 20:28:27
+// Commit: feat(test): add bi-directional event bus handler for test
 
 class testComponentState {
-  final String commitRef = '111';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-04-04T14:17:35+05:00');
+  final String commitRef = '127';
+  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
+  final DateTime timestamp = DateTime.parse('2026-04-04T20:28:27+05:00');
 }